+++ packages/html/init.js
@@ -54,10 +54,8 @@ function init (input, options = {}) {
           textrules.forEach(rule => {
             rule(node)
           })
-          debugger
           // 自定义规则
           textRule && textRule.forEach(rule => {
-            debugger
             rule(node)
           })
         }
