+++ packages/rules/script/chooseRule.js
@@ -19,7 +19,14 @@ var chooseRule = (expression, parentNode, attrKey) => {
     case 'ExpressionStatement': ExpressionStatement(expression); break; // this.xx = 'test'
     case 'CallExpression': CallExpression(expression); break;// cb("a", "中文")
     case 'IfStatement': ; // if (a == "中文")
-    case 'VariableDeclaration': ; // var a 不处理
+    case 'VariableDeclarator':
+      switch(expression.init.type){
+        case 'ObjectExpression':
+          ObjectExpression(expression.init);
+          break;
+
+      }
+      // ObjectExpression(expression.init); // var a 不处理
     default: console.log("notexist:" + expression.type);
   }
 }
